Bugfix and refactoring for Tag deletion.
Previously, only the source of the Tag Annotation would be able to delete it. But now this goes through the BioCatalogue::Auth module to allow curators to delete tags too.

git-svn-id: svn+ssh://rubyforge.org/var/svn/biocatalogue/trunk@1861 1acaf488-1b9d-4964-8077-0a2fdb83ef00

+++ app/controllers/tags_controller.rb
@@ -53,13 +53,14 @@ class TagsController < ApplicationController
                                    :conditions => { :annotatable_type => annotatable.class.name,
                                                     :annotatable_id => annotatable.id,
                                                     :attribute_id => AnnotationAttribute.find_by_name("tag").id, 
-                                                    :source_type => current_user.class.name,
-                                                    :source_id => current_user.id,
                                                     :value => @tag_name })
         
         unless existing.blank?
           existing.each do |a|
-            a.destroy
+            submitters = [ BioCatalogue::Mapper.compound_id_for_model_object(a.source) ]
+            if BioCatalogue::Auth.allow_user_to_curate_thing?(current_user, :tag, :tag_submitters => submitters)            
+              a.destroy
+            end
           end
         end
       end
